[TM-2415] Move the work of formGet to a new service.
