+++ apps/entity-service/src/app.module.ts
@@ -23,6 +23,7 @@ import { DisturbanceService } from "./entities/disturbance.service";
 import { OptionLabelsController } from "./forms/option-labels.controller";
 import { LinkedFieldsController } from "./forms/linked-fields.controller";
 import { FormsController } from "./forms/forms.controller";
+import { FormsService } from "./forms/forms.service";
 
 @Module({
   imports: [SentryModule.forRoot(), CommonModule, HealthModule],
@@ -55,7 +56,8 @@ import { FormsController } from "./forms/forms.controller";
     ImpactStoryService,
     TasksService,
     DemographicService,
-    DisturbanceService
+    DisturbanceService,
+    FormsService
   ]
 })
 export class AppModule {}
